listMajors
